* Started to fix an issue causing lambda variables not to be destroyed properly
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@159 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ cxx/base.cpp
@@ -103,11 +103,11 @@ struct tm xdispatch::as_struct_tm(dispatch_time_t t){
 
     res = *(localtime( &rawtime));
 
-    res.tm_hour += (int)(t / (3600*NSEC_PER_SEC));
-    t %= (3600*NSEC_PER_SEC);
-    res.tm_min += (int)(t / (60*NSEC_PER_SEC));
-    t %= (60*NSEC_PER_SEC);
-    res.tm_sec += (int)(t / (NSEC_PER_SEC));
+    res.tm_hour += (int)(t / (3600*(uint64_t)NSEC_PER_SEC));
+    t %= (3600*(uint64_t)NSEC_PER_SEC);
+    res.tm_min += (int)(t / (60*(uint64_t)NSEC_PER_SEC));
+    t %= (60*(uint64_t)NSEC_PER_SEC);
+    res.tm_sec += (int)(t / ((uint64_t)NSEC_PER_SEC));
 
     return res;
 }
